pyproject.toml
[tool.poetry]
name = "notes-app"
version = "0.1.0"
description = ""
authors = ["seKsuuushka <ksenia_merkulova@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.2"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.6"
pydantic-settings = "^2.5.2"
fastapi-login = "^1.10.1"
python-multipart = "^0.0.9"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
